Set url attribute of a resource at the time of initialization
Url attribute of a resource is valid when the resource is fetched
from remote and shouldn't change once fetched, even if the
connection is reset.

Hence the url is initialized using the connection's url_prefix at
the time it is created. Usually it obeys the format:
  http://(:url_prefix)/(:expanded_uri)

But in the case of ReleaseVariant, it is different and needs to be
handled differently.